Refactor repository
