+++ Core/Helpers/HashHelper.cs
